backend/ai-service/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
gunicorn==21.2.0
python-dotenv==1.0.0

# Database & Data
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2
xgboost==2.0.3
lightgbm==4.1.0

# NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2
sentence-transformers==2.2.2

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
face-recognition==1.3.0
dlib==19.24.2

# Audio Processing
SpeechRecognition==3.10.0
pydub==0.25.1
whisper==1.1.10

# Plagiarism Detection
difflib
python-Levenshtein==0.23.0

# Model Management
mlflow==2.9.2
dvc==3.38.1
onnx==1.15.0
onnxruntime==1.16.3

# Explainable AI
shap==0.44.0
lime==0.2.0.1
eli5==0.13.0

# Bias Detection
fairlearn==0.9.0
aif360==0.5.0

# A/B Testing
scipy==1.11.4
statsmodels==0.14.1

# API & Communication
requests==2.31.0
celery==5.3.4
rabbitpy==2.0.1

# Utilities
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2023.3
joblib==1.3.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
